refactor(perl-parser): split heredoc anti-pattern module by responsibility (#0000)
