+++ crates/perl-parser/src/heredoc_anti_patterns/detectors.rs
@@ -1,89 +1,16 @@
-//! Anti-pattern detection for heredoc edge cases
-//!
-//! This crate provides detection and analysis of problematic Perl patterns
-//! that make static parsing difficult or impossible, particularly around heredocs.
-//!
-//! The [`AntiPatternDetector`] scans Perl source for seven categories of
-//! heredoc-related anti-patterns and produces [`Diagnostic`]s describing each
-//! finding, with severity, explanation, suggested fix, and documentation
-//! references.
-
 use regex::Regex;
 use std::collections::HashSet;
 use std::sync::LazyLock;
 
-/// Source location of a detected anti-pattern.
-///
-/// All three coordinates are provided so callers can serve both LSP (line/column)
-/// and byte-level (offset) consumers without re-computing positions.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Location {
-    /// Zero-based line number within the scanned source fragment.
-    pub line: usize,
-    /// Zero-based column (byte offset from the start of the line).
-    pub column: usize,
-    /// Absolute byte offset from the start of the scanned source fragment.
-    pub offset: usize,
-}
-
-/// Diagnostic severity level for a detected anti-pattern.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Severity {
-    /// The construct will likely cause a runtime or parse failure.
-    Error,
-    /// The construct works but is fragile or difficult to analyze statically.
-    Warning,
-    /// The construct is valid but could be improved for readability or tooling support.
-    Info,
-}
-
-/// A specific category of heredoc-related anti-pattern found in Perl source.
-///
-/// Each variant captures the [`Location`] of the offending construct plus any
-/// context needed to produce a useful diagnostic message.
-#[derive(Debug, Clone, PartialEq)]
-pub enum AntiPattern {
-    /// A heredoc declared inside a `format` body.
-    FormatHeredoc { location: Location, format_name: String, heredoc_delimiter: String },
-    /// A heredoc declared inside a `BEGIN { ... }` block, evaluated at compile time.
-    BeginTimeHeredoc { location: Location, heredoc_content: String, side_effects: Vec<String> },
-    /// A heredoc whose terminator is determined by a variable or expression at runtime.
-    DynamicHeredocDelimiter { location: Location, expression: String },
-    /// A `use Filter::*` statement that may rewrite source before static analysis runs.
-    SourceFilterHeredoc { location: Location, module: String },
-    /// A heredoc embedded inside a `(?{ ... })` regex code block.
-    RegexCodeBlockHeredoc { location: Location },
-    /// A heredoc embedded inside a string argument to `eval`.
-    EvalStringHeredoc { location: Location },
-    /// A heredoc written to a filehandle that has been `tie`d to a custom class.
-    TiedHandleHeredoc { location: Location, handle_name: String },
-}
-
-/// A fully-formed diagnostic produced by the anti-pattern detector.
-///
-/// Contains everything needed to display a problem in an IDE or report:
-/// the severity, the matched pattern (with location), a human-readable message,
-/// a longer explanation, an optional suggested fix, and `perldoc` references.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Diagnostic {
-    /// How serious the problem is.
-    pub severity: Severity,
-    /// The specific anti-pattern that triggered this diagnostic.
-    pub pattern: AntiPattern,
-    /// Short one-line summary suitable for an IDE problem marker.
-    pub message: String,
-    /// Longer explanation of why the construct is problematic.
-    pub explanation: String,
-    /// Optional concrete suggestion for fixing the problem.
-    pub suggested_fix: Option<String>,
-    /// Relevant `perldoc` pages or documentation references.
-    pub references: Vec<String>,
-}
+use crate::heredoc_anti_patterns::model::{AntiPattern, Diagnostic, Location, Severity};
+use crate::heredoc_anti_patterns::utils::{
+    build_line_starts, location_from_start, mask_non_code_regions,
+};
 
 /// Scans Perl source for heredoc-related anti-patterns and produces [`Diagnostic`]s.
 ///
 /// Construct with [`AntiPatternDetector::new`], then call [`detect_all`] with the
-/// source text.  The detector runs all seven built-in pattern checkers and returns
+/// source text. The detector runs all seven built-in pattern checkers and returns
 /// the results sorted by byte offset so callers receive problems in source order.
 ///
 /// [`detect_all`]: AntiPatternDetector::detect_all
@@ -99,96 +26,6 @@ trait PatternDetector: Send + Sync {
         line_starts: &[usize],
     ) -> Vec<(AntiPattern, Location)>;
     fn diagnose(&self, pattern: &AntiPattern) -> Option<Diagnostic>;
-}
-
-fn build_line_starts(code: &str) -> Vec<usize> {
-    let mut line_starts = Vec::new();
-    line_starts.push(0);
-
-    for (idx, byte) in code.bytes().enumerate() {
-        if byte == b'\n' {
-            line_starts.push(idx + 1);
-        }
-    }
-
-    line_starts
-}
-
-fn location_from_start(line_starts: &[usize], offset: usize, start: usize) -> Location {
-    let insertion = line_starts.partition_point(|&line_start| line_start <= start);
-    let line = insertion.saturating_sub(1);
-    let line_start = line_starts.get(line).copied().unwrap_or(0);
-    let column = start.saturating_sub(line_start);
-
-    Location { line, column, offset: offset + start }
-}
-
-fn mask_non_code_regions(code: &str) -> String {
-    fn push_masked_char(masked: &mut String, ch: char) {
-        for _ in 0..ch.len_utf8() {
-            masked.push(' ');
-        }
-    }
-
-    let mut masked = String::with_capacity(code.len());
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-    let mut in_line_comment = false;
-    let mut escaped = false;
-
-    for ch in code.chars() {
-        if in_line_comment {
-            if ch == '\n' {
-                in_line_comment = false;
-                masked.push('\n');
-            } else {
-                push_masked_char(&mut masked, ch);
-            }
-            continue;
-        }
-
-        if in_single_quote {
-            if escaped {
-                escaped = false;
-            } else if ch == '\\' {
-                escaped = true;
-            } else if ch == '\'' {
-                in_single_quote = false;
-            }
-            push_masked_char(&mut masked, ch);
-            continue;
-        }
-
-        if in_double_quote {
-            if escaped {
-                escaped = false;
-            } else if ch == '\\' {
-                escaped = true;
-            } else if ch == '"' {
-                in_double_quote = false;
-            }
-            push_masked_char(&mut masked, ch);
-            continue;
-        }
-
-        match ch {
-            '#' => {
-                in_line_comment = true;
-                push_masked_char(&mut masked, ch);
-            }
-            '\'' => {
-                in_single_quote = true;
-                push_masked_char(&mut masked, ch);
-            }
-            '"' => {
-                in_double_quote = true;
-                push_masked_char(&mut masked, ch);
-            }
-            _ => masked.push(ch),
-        }
-    }
-
-    masked
 }
 
 // Format heredoc detector
@@ -767,308 +604,4 @@ impl AntiPatternDetector {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_format_heredoc_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r#"
-format REPORT =
-<<'END'
-Name: @<<<<<<<<<<<<
-$name
-END
-.
-"#;
-
-        let diagnostics = detector.detect_all(code);
-        // Note: DynamicDelimiterDetector might also flag the << inside the format body as a false positive.
-        // But FormatHeredoc should appear first because it starts at 'format'.
-        // So diagnostics[0] should be FormatHeredoc.
-        assert!(!diagnostics.is_empty());
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::FormatHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_begin_heredoc_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-BEGIN {
-    $config = <<'END';
-    server = localhost
-END
-}
-"###;
-
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::BeginTimeHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_begin_heredoc_detection_with_nested_braces() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-BEGIN {
-    if ($ENV{DEV}) {
-        $config = <<'END';
-        server = localhost
-END
-    }
-}
-"###;
-
-        let diagnostics = detector.detect_all(code);
-        let begin_count = diagnostics
-            .iter()
-            .filter(|diag| matches!(diag.pattern, AntiPattern::BeginTimeHeredoc { .. }))
-            .count();
-        assert_eq!(begin_count, 1);
-    }
-
-    #[test]
-    fn test_dynamic_delimiter_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-my $delimiter = "EOF";
-my $content = <<$delimiter;
-This is dynamic
-EOF
-"###;
-
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::DynamicHeredocDelimiter { .. }));
-    }
-
-    #[test]
-    fn test_source_filter_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-use Filter::Simple;
-print <<EOF;
-Filtered content
-EOF
-"###;
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::SourceFilterHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_regex_heredoc_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-m/pattern(?{
-    print <<'MATCH';
-    Match text
-MATCH
-})/
-"###;
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::RegexCodeBlockHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_eval_heredoc_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-eval 'print <<"EVAL";
-Eval content
-EVAL';
-"###;
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::EvalStringHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_tied_handle_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-tie *FH, 'Tie::Handle';
-print FH <<'DATA';
-Tied output
-DATA
-"###;
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::TiedHandleHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_tied_scalar_handle_detection() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-tie $fh, 'Tie::Handle';
-print $fh <<'DATA';
-Tied output
-DATA
-"###;
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        assert!(matches!(diagnostics[0].pattern, AntiPattern::TiedHandleHeredoc { .. }));
-    }
-
-    #[test]
-    fn test_tied_handle_reports_multiple_writes() {
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-tie *FH, 'Tie::Handle';
-print FH <<'FIRST';
-One
-FIRST
-print FH <<'SECOND';
-Two
-SECOND
-"###;
-
-        let diagnostics = detector.detect_all(code);
-        let tied_handle_count = diagnostics
-            .iter()
-            .filter(|diag| matches!(diag.pattern, AntiPattern::TiedHandleHeredoc { .. }))
-            .count();
-        assert_eq!(tied_handle_count, 2);
-    }
-
-    #[test]
-    fn test_tied_handle_does_not_report_other_handles() {
-        // Regression: PRINT_HEREDOC_PATTERN must only flag handles in the tied set.
-        // Writing a heredoc to an *untied* handle (OTHER) must not produce a diagnostic.
-        let detector = AntiPatternDetector::new();
-        let code = r###"
-tie *FH, 'Tie::Handle';
-print OTHER <<'DATA';
-Not tied
-DATA
-"###;
-
-        let diagnostics = detector.detect_all(code);
-        let tied_handle_count = diagnostics
-            .iter()
-            .filter(|diag| matches!(diag.pattern, AntiPattern::TiedHandleHeredoc { .. }))
-            .count();
-        assert_eq!(tied_handle_count, 0);
-    }
-
-    #[test]
-    fn test_location_column_is_zero_based_for_new_line_matches() {
-        let detector = AntiPatternDetector::new();
-        let code = "my $x = 1;\nuse Filter::Simple;\n";
-
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-
-        assert!(
-            matches!(diagnostics[0].pattern, AntiPattern::SourceFilterHeredoc { .. }),
-            "expected SourceFilterHeredoc pattern, got: {:?}",
-            diagnostics[0].pattern
-        );
-        let AntiPattern::SourceFilterHeredoc { location, .. } = &diagnostics[0].pattern else {
-            return;
-        };
-
-        assert_eq!(location.line, 1);
-        assert_eq!(location.column, 0);
-        assert_eq!(location.offset, 11);
-    }
-
-    #[test]
-    fn test_location_first_byte_is_line_zero_column_zero() {
-        // A match at byte offset 0 must report line=0, column=0.
-        let detector = AntiPatternDetector::new();
-        let code = "use Filter::Simple;\n";
-
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        let AntiPattern::SourceFilterHeredoc { location, .. } = &diagnostics[0].pattern else {
-            unreachable!("expected SourceFilterHeredoc");
-        };
-        assert_eq!(location.line, 0, "first-byte match must be on line 0");
-        assert_eq!(location.column, 0, "first-byte match must be at column 0");
-        assert_eq!(location.offset, 0);
-    }
-
-    #[test]
-    fn test_location_third_line_accurate() {
-        // Three-line file — match on line 2, column 0.
-        let detector = AntiPatternDetector::new();
-        // Line 0: "my $a = 1;\n"  (11 bytes, \n at index 10)
-        // Line 1: "my $b = 2;\n"  (11 bytes, \n at index 21)
-        // Line 2: "use Filter::Simple;\n"
-        let code = "my $a = 1;\nmy $b = 2;\nuse Filter::Simple;\n";
-
-        let diagnostics = detector.detect_all(code);
-        assert_eq!(diagnostics.len(), 1);
-        let AntiPattern::SourceFilterHeredoc { location, .. } = &diagnostics[0].pattern else {
-            unreachable!("expected SourceFilterHeredoc");
-        };
-        assert_eq!(location.line, 2, "match on third line must report line 2");
-        assert_eq!(location.column, 0, "match at start of line must report column 0");
-        assert_eq!(location.offset, 22, "byte offset of third-line start");
-    }
-
-    #[test]
-    fn test_location_mid_line_column_nonzero() {
-        // Match that does not start at column 0 must report the correct column.
-        // Line 0: "# comment\n"      (10 bytes, \n at index 9)
-        // Line 1: "    use Filter::Simple;\n"  — 4 leading spaces, match at column 4
-        let detector = AntiPatternDetector::new();
-        let code = "# comment\n    use Filter::Simple;\n";
-
-        let diagnostics = detector.detect_all(code);
-        // The comment is masked; only SourceFilterHeredoc on line 1 should fire.
-        assert_eq!(diagnostics.len(), 1);
-        let AntiPattern::SourceFilterHeredoc { location, .. } = &diagnostics[0].pattern else {
-            unreachable!("expected SourceFilterHeredoc");
-        };
-        assert_eq!(location.line, 1);
-        assert_eq!(location.column, 4, "mid-line match must report correct column");
-        assert_eq!(location.offset, 14, "byte offset = 10 (first line) + 4 spaces");
-    }
-
-    #[test]
-    fn test_source_filter_detection_ignores_comments_and_strings() {
-        let detector = AntiPatternDetector::new();
-        let code = r#"
-# use Filter::Simple;
-my $s = "use Filter::Simple";
-"#;
-
-        let diagnostics = detector.detect_all(code);
-        assert!(diagnostics.is_empty());
-    }
-
-    #[test]
-    fn test_begin_detection_ignores_comments_and_strings() {
-        let detector = AntiPatternDetector::new();
-        let code = r#"
-# BEGIN { my $x = <<'END'; END }
-my $s = "BEGIN { my $x = <<'END'; END }";
-"#;
-
-        let diagnostics = detector.detect_all(code);
-        assert!(diagnostics.is_empty());
-    }
-
-    #[test]
-    fn test_format_detection_handles_utf8_in_masked_regions() {
-        let detector = AntiPatternDetector::new();
-        let code = r#"# comment with emoji 😀
-format REPORT =
-<<'END'
-Body
-END
-.
-"#;
-
-        let diagnostics = detector.detect_all(code);
-        assert!(
-            diagnostics
-                .iter()
-                .any(|diag| matches!(diag.pattern, AntiPattern::FormatHeredoc { .. }))
-        );
-    }
-}
+mod tests;
